Make Integer#zero? a separated method and builtin (#3226)
A prerequisite to fix https://bugs.ruby-lang.org/issues/15589 with JIT.
This commit alone doesn't make a significant difference yet, but I thought
this commit should be committed independently.

This method override was discussed in [Misc #16961].

+++ test/ruby/test_jit.rb
@@ -980,7 +980,7 @@ class TestJIT < Test::Unit::TestCase
   def test_frame_omitted_inlining
     assert_eval_with_jit("#{<<~"begin;"}\n#{<<~"end;"}", stdout: "true\ntrue\ntrue\n", success_count: 1, min_calls: 2)
     begin;
-      class Numeric
+      class Integer
         remove_method :zero?
         def zero?
           self == 0
